fix: keep review output adapter focused
